feat: enable local builds of wordpress plugin zip via make
The only way to create the commonsbooking.zip Wordpress Plugin archive
locally was by calling the zipdist.sh script. Unfortunately, this script
doesn’t contain any logic on how to build the actual plugin but knows
only which directories and files to zip.

This made reliable and somewhat reproducible builds from experimental
code branches difficult.

This commit introduces a Makefile with the following targets:

  * build
    Builds the commonsbooking plugin archive in the local environment.
  * build-in-docker
    Builds the commonsbooking plugin archive in a docker container,
    therefore isolating the build process.
  * clean
    Removes any built, generated, or downloaded files.

The targets can be called with `make TARGET_NAME`.