TimeframeExport: Fixes #1532
Adds the fix and additional tests to validate the fix.

The commit 8d98c00317e219dc0185adab8ec1e9f88ebba949 introduced this bug.
It had the purpose to allow timesframes (for holidays) to be "mutli-selected" with more than one location or item.
This fixes the export, but won't include custom-fields of location- and item-posts, related to holiday-timeframes in the
export file.