fix BookingCodesMessage:sendMessage() wrong $attachment for prepareMail
prepareMail expects an associative array or null as $attachment, but sendMessage passed a associate array inside a array (which was actually working due to magic in Message::addStringAttachments) or null inside an array, for the case where an ical-attachment is not to be sent. The [null] was misunderstood by SendNotificationMail() as a mail with an attachment, calling phpmailer with illegal parameters and causing warnings. The mail was seemingly send anyway, totally expected without attachment. So the change is a fix to avoid warnings, but it does not change the behavior for the user in any way.

Additionally, add two more tests to cover the (non-default) case when sending of ical-attachment for BookingCodes is ON