Fix #1712 with explicit sass imports via use
The migration of sass in commit 37059c40da90fea0e03ea1170b3d073a918e0a47 introduced several occurences
of the optional keyword. This caused the styles being uncompletly generated, because the class-names
before the keyword did not get resolved properly and thus CSS not being generated.
This was because I did not understood the problem enough.
This commit introduces explicit commits to the partials files that define the class-names.