make getEndSlot use timeframe model methods, fix wonky tests caused by unexpected grid output
When setting up timeframes in the unit tests, the fullday meta field gets set to "off" while in the real world, CMB2 empties this field when the checkbox is not set. getEndSlot did not previously use the method on the Timeframe model and therefore would falsly assume, that a timeframe which has the fullday field set to "off" has a value and therefore must have fullday enabled.